+++ src/main/java/com/matchFit/participation/repository/ParticipationRepository.java
@@ -1,6 +1,8 @@
-package com.matchFit.participation.entity;
+package com.matchFit.participation.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.matchFit.participation.entity.Participation;
 
 public interface ParticipationRepository extends JpaRepository<Participation, Long> {
 	// 현재 신청자 수 계산
