+++ src/main/java/com/matchFit/post/repository/PostRepository.java
@@ -1,4 +1,4 @@
-package com.matchFit.post;
+package com.matchFit.post.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
